third_party/blink/web_tests/platform/mac-mac15-arm64/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/log.https.any.worker_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] log float16 positive 1D tensor
  assert_array_approx_equals: test log float16 property 3, expected 3.7109375 +/- 0.0009765625, expected 3.7109375 but got 3.712890625
[FAIL] log float16 positive 2D tensor
  assert_array_approx_equals: test log float16 property 3, expected 3.7109375 +/- 0.0009765625, expected 3.7109375 but got 3.712890625
[FAIL] log float16 positive 3D tensor
  assert_array_approx_equals: test log float16 property 3, expected 3.7109375 +/- 0.0009765625, expected 3.7109375 but got 3.712890625
[FAIL] log float16 positive 4D tensor
  assert_array_approx_equals: test log float16 property 3, expected 3.7109375 +/- 0.0009765625, expected 3.7109375 but got 3.712890625
[FAIL] log float16 positive 5D tensor
  assert_array_approx_equals: test log float16 property 3, expected 3.7109375 +/- 0.0009765625, expected 3.7109375 but got 3.712890625
Harness: the test ran to completion.
